Extract knowledge file listing into knowledge/file_listing.py
Move the file listing and inclusion subsystem out of knowledge/manager.py
into a focused module with one stated contract: every returned path is a
regular file, not a symlink, with no symlinked ancestors, strictly
resolved inside the knowledge root.

The extraction also untangles the layering:
- Traversal (_iter_target_files) owns directory-chain vetting via a
  cached _DirectoryGuard, so the check_directory_symlinks flag and its
  non-local safety invariant are gone.
- Per-file safety checks (_resolve_safe_file) return the resolved path,
  removing the duplicate resolve() previously paid per accepted file and
  letting dedup fall out of a plain set.
- Entry points compose pattern rules, guard, and safety checks cheap to
  expensive at each call site.
