+++ src/mindroom/knowledge/__init__.py
@@ -3,7 +3,7 @@
 # ruff: noqa: RUF022
 
 from mindroom.knowledge.availability import KnowledgeAvailability
-from mindroom.knowledge.manager import list_knowledge_files
+from mindroom.knowledge.file_listing import list_knowledge_files
 from mindroom.knowledge.refresh_scheduler import KnowledgeRefreshScheduler
 from mindroom.knowledge.status import reconcile_knowledge_mode_transition_states
 from mindroom.knowledge.utils import (
